agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]


[project]
name = "accomps"
version = "0.1.0"
description = "Accomplishment tracking agent using LangChain"
authors = [{ name = "Shargrove09", email = "8881163+Shargrove09@users.noreply.github.com " }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "langchain>=1.0.5",
    "langchain-openai>=0.2.14",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]
